third_party/blink/web_tests/traversal/moz-mutations-3-expected.txt

CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root B D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B [E] * E1 E11 (actual)
CONSOLE MESSAGE: root B [E] * E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root E [E1] * E11 (actual)
CONSOLE MESSAGE: root E [E1] * E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root E E1 [E11] * (actual)
CONSOLE MESSAGE: root E E1 [E11] * (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root E E1 [E11] * (actual)
CONSOLE MESSAGE: root E E1 [E11] * (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root * [E] E1 E11 (actual)
CONSOLE MESSAGE: root * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B C D E [E1] * E11 (actual)
CONSOLE MESSAGE: root B C D E [E1] * E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C * [E] E1 E11 (actual)
CONSOLE MESSAGE: root C * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C n * [E] E1 E11 (actual)
CONSOLE MESSAGE: root C n * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root n2 C n * [E] E1 E11 (actual)
CONSOLE MESSAGE: root n2 C n * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root B C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root B C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C D * [E] E1 E11 (actual)
CONSOLE MESSAGE: root C D * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C * [E] E1 E11 (actual)
CONSOLE MESSAGE: root C * [E] E1 E11 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C E E1 [E11] * (actual)
CONSOLE MESSAGE: root C E E1 [E11] * (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C [E] * (actual)
CONSOLE MESSAGE: root C [E] * (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C n [E] * (actual)
CONSOLE MESSAGE: root C n [E] * (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C n [E] * n2 (actual)
CONSOLE MESSAGE: root C n [E] * n2 (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C n E [n2] * (actual)
CONSOLE MESSAGE: root C n E [n2] * (expected)
CONSOLE MESSAGE: sequence check:
CONSOLE MESSAGE: root C n (actual)
CONSOLE MESSAGE: root C n (expected)
Results available in the JavaScript console
Test adapted from http://mxr.mozilla.org/mozilla-central/source/content/base/test/test_NodeIterator_mutations_3.html
